fix(examples): examples now works
This isn't a very good example script, but at least it works without completely
crashing. It still has the problem where it crashes at the end of Python
process, but we've got a work item for that.

However, this isn't the final step. There was a stupid hack that I used to get
things going on line 177 of src/pyexiv2/xmp.py. Basically, we were getting
crashes because the key and value for each of the items in the map being passed
to self._tag._setLangAltValue were bytes. This forces them into strings. I know
there's a lot of work in this library to handle different encodings, so I
should make sure that we can't fix that issue earlier in the process,
especially as they still show up as bytes when loaded later.

Part of pridkett/pyexiv2#7

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>